% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gsm.R
\name{gsm_design}
\alias{gsm_design}
\title{Extract design information from an stpm2/gsm object and newdata
for use in C++}
\usage{
gsm_design(object, newdata, newdata0 = NULL, t0 = NULL, inflate = 100)
}
\arguments{
\item{object}{stpm2/gsm object}

\item{newdata}{list or data-frame used for evaluation}

\item{newdata0}{list or data-frame used for evaluation at the entry time}

\item{t0}{possible delayed entry time (numeric scalar)}

\item{inflate}{double value to inflate minimum and maximum times for root finding}
}
\value{
list that can be read by `gsm ssim::read_gsm(SEX args)` in C++
}
\description{
Extract design information from an stpm2/gsm object and newdata
for use in C++
}
