\name{lines-methods}
\docType{methods}
\alias{lines-methods}
\alias{lines,ANY-method}
\title{ lines generic method}
\description{
provides a lines generic method
}
\section{Methods}{
\describe{
\item{\code{signature(x = "ANY")}}{
  generic
}
}}

\keyword{methods}
%%\keyword{ ~~ other possible keyword(s)}


