\name{update-methods}
\docType{methods}
\alias{update-methods}
\alias{update,stpm2-method}
\title{ Methods for Function update}
\description{
 Methods for function \code{update}
}
\section{Methods}{
  \describe{

    \item{update}{\code{signature(object = "stpm2", ...)}: Similar to
      update.default, using S4 methods. }
  }
}
\keyword{methods}
%% \keyword{ ~~ other possible keyword(s)}
